Module fil_actor_power_state::v14

Structs

Claim

CreateMinerParams

CreateMinerReturn

CronEvent

CurrentTotalPowerReturn

EnrollCronEventParams

INITIAL_QA_POWER_ESTIMATE_POSITION

genesis power in bytes = 750,000 GiB

INITIAL_QA_POWER_ESTIMATE_VELOCITY

max chain throughput in bytes per epoch = 120 ProveCommits / epoch = 3,840 GiB

MinerConsensusCountReturn

MinerCountReturn

MinerRawPowerParams

MinerRawPowerReturn

NetworkRawPowerReturn

State

Storage power actor state

SubmitPoRepForBulkVerifyParams

UpdateClaimedPowerParams

UpdatePledgeTotalParams

Enums

Method

Storage power actor methods available

Constants

CLAIMS_CONFIG

CONSENSUS_MINER_MIN_MINERS

Minimum power of an individual miner to meet the threshold for leader election.

CRON_QUEUE_AMT_BITWIDTH

CRON_QUEUE_HAMT_BITWIDTH

ERR_TOO_MANY_PROVE_COMMITS

MAX_MINER_PROVE_COMMITS_PER_EPOCH

Maximum number of prove commits a miner can submit in one epoch

SECTOR_TERMINATION_EXPIRED

Implicit termination after all deals expire

SECTOR_TERMINATION_FAULTY

Implicit termination due to unrecovered fault

SECTOR_TERMINATION_MANUAL

Unscheduled explicit termination by the miner

Functions

consensus_miner_min_power

Returns the minimum storage power required for each PoSt proof type.

epoch_key

set_claim

Type Aliases

ClaimsMap

SectorTermination